requirements.txt
Django==3.2.8
djangorestframework-simplejwt==4.8.0
psycopg2==2.9.1
PyJWT==2.1.0
django-phonenumber-field[phonenumberslite]